# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vastora Center API",
    "version": "v1",
    "summary": "Administrator and Agent control-plane contract",
    "description": "The current prerelease /api/v1 contract. Browser administrator reads require the SameSite session cookie; mutations additionally require the X-CSRF-Token header. Agent endpoints use a per-Agent bearer credential, initial binary download uses a one-time enrollment bearer token, and host-removal completion uses a token bound to one decommission task. JSON requests must use application/json, contain one value, be at most 1 MiB, and contain no unknown fields. Error responses use the Error envelope and never include secrets. Binary downloads and event streams declare their actual media types."
  },
  "servers": [
    {
      "url": "/",
      "description": "The current Center"
    }
  ],
...
  "paths": {
    "/api/v1/agent-binaries/{os}/{arch}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Agent Binary",
        "operationId": "agentBinary_get",
        "x-vastora-audience": "agent-enrollment",
        "security": [
          {
            "EnrollmentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "os",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "arch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ]
      }
    },
    "/api/v1/agent-decommission-results/{taskID}": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Complete Agent Decommission Callback",
        "operationId": "completeAgentDecommissionCallback_post",
        "x-vastora-audience": "agent-decommission-callback",
        "security": [
          {
            "DecommissionCallbackBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "taskID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "A single strict JSON object. Content-Type must be application/json, the decoded size is limited to 1 MiB, and unknown fields are rejected.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "attempt": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/{id}/binary/{os}/{arch}": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Agent Update Binary",
        "operationId": "agentUpdateBinary_get",
        "x-vastora-audience": "agent",
        "security": [
          {
            "AgentBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "os",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "arch",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ]
      }
    },
    "/api/v1/setup/status": {
      "get": {
        "tags": [
          "Setup"
        ],
        "summary": "Setup Status",
        "operationId": "setupStatus_get",
        "x-vastora-audience": "bootstrap-optional-admin",
        "security": [
          {},
          {
            "AdminSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/setup/admin": {
      "post": {
        "tags": [
          "Setup"
        ],
        "summary": "Setup Admin",
        "operationId": "setupAdmin_post",
        "x-vastora-audience": "bootstrap-public",
        "security": [],
        "responses": {
          "201": {
            "description": "Created.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "description": "A single strict JSON object. Content-Type must be application/json, the decoded size is limited to 1 MiB, and unknown fields are rejected.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_post",
        "x-vastora-audience": "bootstrap-public",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "description": "The credentials were rejected.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginError"
                }
              }
            }
          },
          "403": {
            "description": "The required login security check failed or is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginError"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "description": "Sign-in is temporarily unavailable. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "description": "A single strict JSON object. Content-Type must be application/json, the decoded size is limited to 1 MiB, and unknown fields are rejected.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "turnstileToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "description": "Authenticates the administrator with server-enforced abuse protection. Public responses do not expose failure thresholds, lockout durations or retry countdowns. Complete the security check when captchaRequired is true."
      }
    },
...
    "/api/v1/catalog/official": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Official Catalog",
        "operationId": "officialCatalog_get",
        "x-vastora-audience": "bootstrap-public",
        "security": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vastora_session",
        "description": "Administrator session cookie."
      },
      "AdminCSRF": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required together with AdminSession for browser mutations."
      },
      "AgentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque Agent credential",
        "description": "Credential bound to the Agent id in the path."
      },
      "EnrollmentBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "one-time enrollment token",
        "description": "Short-lived token used only to download the initial Agent binary."
      },
      "DecommissionCallbackBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "task-bound callback token",
        "description": "Single-task token used only to acknowledge completed local Agent removal."
      }
    },
...
  }
}